pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xtractor"
version = "0.1.0"
description = "Document extraction pipeline with LangGraph multi-agent orchestration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "langchain>=0.2.14",
    "langchain-core>=0.3.76",
    "langchain-openai>=0.3.33",
    "langgraph>=0.0.70",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.2.1",
    "pytest>=8.4.2",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.9",
    "typing-extensions>=4.12.0",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
    "black>=24.3.0",
]

[project.scripts]
xtractor-cli = "xtractor.cli.main:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["xtractor*"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B"]

[tool.black]
line-length = 100
